+++ library/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the library will be documented in this file.
 ## vX.X.X (Month DD, YYYY)
 
 - Add `file`, `function` and `promise` schema
+- Add `awaitAsync` action to await promise in pipeline
 - Add `operation` property to `filterItems`, `findItem`, `mapItems`, `reduceItems` and `sortItem` action
 - Rename `action` argument of `filterItems`, `findItem`, `mapItems`, `reduceItems` and `sortItem` action to `operation`
 - Rename `action` argument and property of `transform` and `transformAsync` action to `operation`