+++ library/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the library will be documented in this file.
 
-## vX.X.X (Month DD, YYYY)
+## v0.34.0 (June 24, 2024)
 
 - Add `file`, `function` and `promise` schema
 - Add `awaitAsync` action to await promise in pipeline